+++ AnkiDroid/src/main/java/com/ichi2/anki/dialogs/customstudy/CustomStudyDialog.java
@@ -269,7 +269,7 @@ public class CustomStudyDialog extends AnalyticsDialogFragment implements
                         }
                         case CUSTOM_STUDY_FORGOT: {
                             JSONArray ar = new JSONArray();
-                            ar.put(0, 1);
+                            ar.put((Integer) 0, 1);
                             createCustomStudySession(ar, new Object[] {String.format(Locale.US,
                                     "rated:%d:1", n), Consts.DYN_MAX_SIZE, Consts.DYN_RANDOM}, false);
                             break;
@@ -519,7 +519,7 @@ public class CustomStudyDialog extends AnalyticsDialogFragment implements
         ar.getJSONArray(0).put(0, "deck:\"" + deckToStudyName + "\" " + terms[0]);
         ar.getJSONArray(0).put(1, terms[1]);
         @Consts.DYN_PRIORITY int priority = (Integer) terms[2];
-        ar.getJSONArray(0).put(2, priority);
+        ar.getJSONArray(0).put((Integer) 2, priority);
         dyn.put("resched", resched);
         // Rebuild the filtered deck
         Timber.i("Rebuilding Custom Study Deck");
